Two conflicts, both real overlaps rather than mechanical ones.

1. crates/trading/src/candles/aggregator_cell.rs — fold into an OPEN bucket

   Both branches fixed this block, and on `oi` they fixed DIFFERENT HALVES of
   the same bug. Resolution keeps both guards rather than picking a side:

   - close / close_ts: main's ORDER GUARD is taken as-is and supersedes this
     branch's unconditional last-write-wins. Main's is strictly stronger — an
     out-of-order packet inside a still-open bucket could rewrite the close,
     and on the daily bar that window is the whole session.

   - oi: BOTH guards, nested. The order guard answers "is this packet newer?";
     it cannot answer "does this packet carry an OI reading at all". `0` is the
     ABSENT sentinel — a Ticker-mode packet has no OI field and an equity never
     has one — so a NEWER blank packet passes the order guard and would still
     erase a real OI established earlier in the same bucket. Order alone does
     not turn a blank field into news. Last-non-zero now applies INSIDE the
     order guard, with the ignored-zero case counted.

   - volume: main's `.max()` semantics kept, written as an explicit compare so
     the suppressed case increments tv_candle_volume_regression_suppressed_total.
     The value is identical either way; the difference is that a silent `.max()`
     tells you nothing about how often reordering actually happens, and that
     rate is what says whether the guard is load-bearing or dormant.

2. .claude/plans/active-plan-feed-hardening.md — colliding item numbers

   main and this branch each appended an "Item 12" on consecutive days, to
   different work. Both kept, in date order, with a merge note naming the
   collision and the reading rule. Deliberately NOT renumbered: the body text
   in each block cross-references its own item numbers, so renumbering the
   headings alone would leave every reference pointing at the wrong section —
   worse than the collision it would fix.

Verification after merge (measured):
  core lib     2323 passed, 0 failed
  storage lib   926 passed, 0 failed
  trading lib  1621 passed, 0 failed  (up from 1597 — main's 24 new tests all
                                       pass against this resolution)
  app lib      1372 passed, 0 failed
  api lib       332 passed, 0 failed
  cargo fmt --check clean; clippy -D warnings -W clippy::perf clean

+++ crates/app/tests/dhat_live_ingest_seam.rs
@@ -151,7 +151,7 @@ fn ingest_tick_seam_does_not_allocate_per_tick() {
     // would make this gate fail for the wrong reason and teach the next reader
     // to raise the budget instead of finding the leak.
     for i in 0..4u64 {
-        let _ = ingest.ingest_tick_at(&tick(13 + i, 100.0, SESSION_EPOCH_SECS), i as u64, 0, 1_000);
+        let _ = ingest.ingest_tick_at(&tick(13 + i, 100.0, SESSION_EPOCH_SECS), i, 0, 1_000);
     }
 
     let profiler = dhat::Profiler::builder().testing().build();
